requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
sentence-transformers
networkx
numpy
scipy
jsonlines